+++ CHANGES.rst
@@ -46,6 +46,8 @@ Bug fixes
 * ``obs_suite``: fixing observation suite level1e tests (:pull:`17`)
 * ``obs_suite``: QC mapping from ``v7.0.0`` is now running by setting values of ``location_quality`` and ``report_time_quality`` to ``str`` (:pull:`18`)
 * ``obs_suite``: use only observation reports that are also available in the header file (:pull:`44`, :pull:`45`)
+* ``qc_suite``: take qc source data from level1d files instead of level1a (:pull:`47`)
+* ``qc_suite``: update deck list for quality control (:pull:`47`)
 
 v7.0.0 (2024-06-13)
 -------------------